chore: move avatar component
