+++ packages/components/avatar/index.js
@@ -1,3 +1,3 @@
-export { default } from '../../../src/components/avatar';
+export { default } from './src';
 
 export { default as version } from './version';
