+++ packages/components/avatar/src/README.md
