+++ packages/components/avatar/src/avatar.js
