+++ packages/components/avatar/src/avatar.story.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text, boolean, select } from '@storybook/addon-knobs/react';
-import Section from '../../../.storybook/decorators/section';
+import Section from '../../../../.storybook/decorators/section';
 import Avatar from './avatar';
 import Readme from './README.md';
