+++ packages/components/avatar/src/avatar.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Avatar } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../test/percy';
+import { Suite, Spec } from '../../../../test/percy';
 
 export const routePath = '/avatar';
