+++ packages/components/avatar/src/avatar.visualspec.js
