+++ packages/components/avatar/src/index.js
